reorg new tests from master
