+++ sdk/core/core/test/cmocka/CMakeLists.txt
@@ -19,15 +19,15 @@ endif()
 
 add_cmocka_test(${TARGET_NAME} SOURCES
                 main.c
-                test_az_json.c
-                test_az_span.c
-                test_az_context.c
-                test_az_logging.c
-                test_az_http.c
-                test_az_pipeline.c
-                test_az_encode.c
                 test_az_aad.c
+                test_az_context.c
+                test_az_encode.c
+                test_az_http.c
+                test_az_json.c
+                test_az_logging.c
+                test_az_pipeline.c
                 test_az_policy.c
+                test_az_span.c
                 COMPILE_OPTIONS ${DEFAULT_C_COMPILE_FLAGS}
                 LINK_OPTIONS ${WRAP_FUNCTIONS}
                 # grant access to Private functions to test az_json_private
