+++ sdk/core/core/test/cmocka/test_az_span.c
@@ -35,7 +35,7 @@ static void az_single_char_ascii_lower_test(void** state)
   for (uint8_t i = 0; i <= SCHAR_MAX; ++i)
   {
     uint8_t buffer[1] = { i };
-    az_span span = AZ_SPAN_LITERAL_FROM_INITIALIZED_BUFFER(buffer);
+    az_span span = AZ_SPAN_FROM_BUFFER(buffer);
 
     // Comparison to itself should return true for all values in the range.
     assert_true(az_span_is_content_equal_ignoring_case(span, span));
@@ -44,14 +44,14 @@ static void az_single_char_ascii_lower_test(void** state)
     if (i >= 'A' && i <= 'Z')
     {
       uint8_t lower[1] = { (uint8_t)(i + 32) };
-      az_span lowerSpan = AZ_SPAN_LITERAL_FROM_INITIALIZED_BUFFER(lower);
+      az_span lowerSpan = AZ_SPAN_FROM_BUFFER(lower);
       assert_true(az_span_is_content_equal_ignoring_case(span, lowerSpan));
       assert_true(az_span_is_content_equal_ignoring_case(lowerSpan, span));
     }
     else if (i >= 'a' && i <= 'z')
     {
       uint8_t upper[1] = { (uint8_t)(i - 32) };
-      az_span upperSpan = AZ_SPAN_LITERAL_FROM_INITIALIZED_BUFFER(upper);
+      az_span upperSpan = AZ_SPAN_FROM_BUFFER(upper);
       assert_true(az_span_is_content_equal_ignoring_case(span, upperSpan));
       assert_true(az_span_is_content_equal_ignoring_case(upperSpan, span));
     }
@@ -61,7 +61,7 @@ static void az_single_char_ascii_lower_test(void** state)
       for (uint8_t j = 0; j <= SCHAR_MAX; ++j)
       {
         uint8_t other[1] = { j };
-        az_span otherSpan = AZ_SPAN_LITERAL_FROM_INITIALIZED_BUFFER(other);
+        az_span otherSpan = AZ_SPAN_FROM_BUFFER(other);
 
         if (i == j)
         {
@@ -74,172 +74,6 @@ static void az_single_char_ascii_lower_test(void** state)
       }
     }
   }
-}
-
-static void az_span_append_uint8_overflow_fails(void** state)
-{
-  (void)state;
-
-  uint8_t raw_buffer[2];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-
-  assert_true(az_succeeded(az_span_append_uint8(buffer, 'a', &buffer)));
-  assert_true(az_succeeded(az_span_append_uint8(buffer, 'b', &buffer)));
-  assert_true(az_failed(az_span_append_uint8(buffer, 'c', &buffer)));
-}
-
-static void az_span_append_uint8_succeeds(void** state)
-{
-  (void)state;
-
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-
-  assert_true(az_succeeded(az_span_append_uint8(buffer, 'a', &buffer)));
-  assert_true(az_succeeded(az_span_append_uint8(buffer, 'b', &buffer)));
-  assert_true(az_succeeded(az_span_append_uint8(buffer, 'c', &buffer)));
-  assert_true(az_span_is_content_equal(buffer, AZ_SPAN_FROM_STR("abc")));
-}
-
-static void az_span_append_i32toa_succeeds(void** state)
-{
-  (void)state;
-
-  int32_t v = 12345;
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_succeeded(az_span_append_i32toa(buffer, v, &out_span)));
-  assert_true(az_span_is_content_equal(out_span, AZ_SPAN_FROM_STR("12345")));
-}
-
-static void az_span_append_i32toa_negative_succeeds(void** state)
-{
-  (void)state;
-
-  int32_t v = -12345;
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_succeeded(az_span_append_i32toa(buffer, v, &out_span)));
-  assert_true(az_span_is_content_equal(out_span, AZ_SPAN_FROM_STR("-12345")));
-}
-
-static void az_span_append_i32toa_zero_succeeds(void** state)
-{
-  (void)state;
-
-  int32_t v = 0;
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_succeeded(az_span_append_i32toa(buffer, v, &out_span)));
-  assert_true(az_span_is_content_equal(out_span, AZ_SPAN_FROM_STR("0")));
-}
-
-static void az_span_append_i32toa_max_int_succeeds(void** state)
-{
-  (void)state;
-
-  int32_t v = 2147483647;
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_succeeded(az_span_append_i32toa(buffer, v, &out_span)));
-  assert_true(az_span_is_content_equal(out_span, AZ_SPAN_FROM_STR("2147483647")));
-}
-
-static void az_span_append_i32toa_NULL_span_fails(void** state)
-{
-  (void)state;
-
-  int32_t v = 2147483647;
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_succeeded(az_span_append_i32toa(buffer, v, &out_span)));
-  assert_true(az_span_is_content_equal(out_span, AZ_SPAN_FROM_STR("2147483647")));
-}
-
-static void az_span_append_i32toa_overflow_fails(void** state)
-{
-  (void)state;
-
-  int32_t v = 2147483647;
-  uint8_t raw_buffer[4];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_span_append_i32toa(buffer, v, &out_span) == AZ_ERROR_INSUFFICIENT_SPAN_CAPACITY);
-}
-
-static void az_span_append_u32toa_succeeds(void** state)
-{
-  (void)state;
-
-  uint32_t v = 12345;
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_succeeded(az_span_append_u32toa(buffer, v, &out_span)));
-  assert_true(az_span_is_content_equal(out_span, AZ_SPAN_FROM_STR("12345")));
-}
-
-static void az_span_append_u32toa_zero_succeeds(void** state)
-{
-  (void)state;
-
-  uint32_t v = 0;
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_succeeded(az_span_append_u32toa(buffer, v, &out_span)));
-  assert_true(az_span_is_content_equal(out_span, AZ_SPAN_FROM_STR("0")));
-}
-
-static void az_span_append_u32toa_max_uint_succeeds(void** state)
-{
-  (void)state;
-
-  uint32_t v = 4294967295;
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_succeeded(az_span_append_u32toa(buffer, v, &out_span)));
-  assert_true(az_span_is_content_equal(out_span, AZ_SPAN_FROM_STR("4294967295")));
-}
-
-static void az_span_append_u32toa_NULL_span_fails(void** state)
-{
-  (void)state;
-
-  uint32_t v = 2147483647;
-  uint8_t raw_buffer[15];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_succeeded(az_span_append_u32toa(buffer, v, &out_span)));
-  assert_true(az_span_is_content_equal(out_span, AZ_SPAN_FROM_STR("2147483647")));
-}
-
-static void az_span_append_u32toa_overflow_fails(void** state)
-{
-  (void)state;
-
-  uint32_t v = 2147483647;
-  uint8_t raw_buffer[4];
-  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
-  az_span out_span;
-
-  assert_true(az_span_append_u32toa(buffer, v, &out_span) == AZ_ERROR_INSUFFICIENT_SPAN_CAPACITY);
 }
 
 static void az_span_to_lower_test(void** state)
@@ -255,36 +89,33 @@ static void az_span_to_lower_test(void** state)
   assert_false(az_span_is_content_equal_ignoring_case(a, d));
 }
 
-static void az_span_to_uint64_return_errors(void** state)
+static void az_span_atou64_return_errors(void** state)
 {
   (void)state;
-
   // sample span
   az_span sample = AZ_SPAN_FROM_STR("test");
   uint64_t out = 0;
 
-  assert_true(az_span_to_uint64(sample, &out) == AZ_ERROR_PARSER_UNEXPECTED_CHAR);
+  assert_true(az_span_atou64(sample, &out) == AZ_ERROR_PARSER_UNEXPECTED_CHAR);
 }
 
-static void az_span_to_uint32_test(void** state)
+static void az_span_atou32_test(void** state)
 {
   (void)state;
-
   az_span number = AZ_SPAN_FROM_STR("1024");
   uint32_t value = 0;
 
-  assert_return_code(az_span_to_uint32(number, &value), AZ_OK);
+  assert_return_code(az_span_atou32(number, &value), AZ_OK);
   assert_int_equal(value, 1024);
 }
 
 static void az_span_to_str_test(void** state)
 {
   (void)state;
-
   az_span sample = AZ_SPAN_FROM_STR("hello World!");
   char str[20];
 
-  assert_return_code(az_span_to_str(str, 20, sample), AZ_OK);
+  az_span_to_str(str, 20, sample);
   assert_string_equal(str, "hello World!");
 }
 
@@ -406,10 +237,15 @@ static void az_span_find_capacity_checks_success(void** state)
 
   uint8_t* buffer = (uint8_t*)"aaaa";
 
-  assert_int_equal(az_span_find(az_span_init(buffer, 2, 4), az_span_init(buffer, 2, 3)), 0);
-  assert_int_equal(az_span_find(az_span_init(buffer, 2, 3), az_span_init(buffer, 2, 4)), 0);
-  assert_int_equal(az_span_find(az_span_init(buffer, 2, 3), az_span_init(buffer, 0, 2)), 0);
-  assert_int_equal(az_span_find(az_span_init(buffer, 0, 2), az_span_init(buffer, 0, 2)), 0);
+  assert_int_equal(az_span_find(az_span_init(buffer, 2), az_span_init(buffer, 2)), 0);
+  assert_int_equal(az_span_find(az_span_init(buffer, 2), az_span_init(buffer, 0)), 0);
+  assert_int_equal(az_span_find(az_span_init(buffer, 0), az_span_init(buffer, 0)), 0);
+
+  assert_int_equal(az_span_find(az_span_init(buffer, 2), az_span_init(buffer + 1, 2)), 0);
+  assert_int_equal(az_span_find(az_span_init(buffer + 1, 2), az_span_init(buffer, 2)), 0);
+  assert_int_equal(az_span_find(az_span_init(buffer + 1, 2), az_span_init(buffer + 1, 2)), 0);
+  assert_int_equal(az_span_find(az_span_init(buffer, 2), az_span_init(buffer + 2, 2)), 0);
+  assert_int_equal(az_span_find(az_span_init(buffer + 2, 2), az_span_init(buffer, 2)), 0);
 }
 
 static void az_span_find_overlapping_checks_success(void** state)
@@ -433,11 +269,16 @@ static void test_az_span_replace(void** state)
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("12345678");
     az_span expected = AZ_SPAN_FROM_STR("1X78");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
 
-    assert_true(_az_span_replace(&builder, 1, 6, AZ_SPAN_FROM_STR("X")) == AZ_OK);
+    int32_t expected_remainder_size = 200 - az_span_size(initial_state);
+    assert_int_equal(az_span_size(remainder), expected_remainder_size);
 
-    az_span const result = builder;
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 1, 6, AZ_SPAN_FROM_STR("X"))
+        == AZ_OK);
+
+    az_span const result = az_span_slice(builder, 0, 4);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -446,11 +287,14 @@ static void test_az_span_replace(void** state)
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("12345678");
     az_span expected = AZ_SPAN_FROM_STR("12X345678");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 200 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 2, 2, AZ_SPAN_FROM_STR("X")) == AZ_OK);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 2, 2, AZ_SPAN_FROM_STR("X"))
+        == AZ_OK);
 
-    az_span const result = builder;
+    az_span const result = az_span_slice(builder, 0, 9);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -459,11 +303,14 @@ static void test_az_span_replace(void** state)
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("12345678");
     az_span expected = AZ_SPAN_FROM_STR("1234567890");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 200 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 8, 8, AZ_SPAN_FROM_STR("90")) == AZ_OK);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 8, 8, AZ_SPAN_FROM_STR("90"))
+        == AZ_OK);
 
-    az_span const result = builder;
+    az_span const result = az_span_slice(builder, 0, 10);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -472,11 +319,14 @@ static void test_az_span_replace(void** state)
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("12345678");
     az_span expected = AZ_SPAN_FROM_STR("X");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 200 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 0, 8, AZ_SPAN_FROM_STR("X")) == AZ_OK);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 0, 8, AZ_SPAN_FROM_STR("X"))
+        == AZ_OK);
 
-    az_span const result = builder;
+    az_span const result = az_span_slice(builder, 0, 1);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -485,11 +335,14 @@ static void test_az_span_replace(void** state)
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("12345678");
     az_span expected = AZ_SPAN_FROM_STR("X12345678X");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 200 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 0, 8, AZ_SPAN_FROM_STR("X12345678X")) == AZ_OK);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 0, 8, AZ_SPAN_FROM_STR("X12345678X"))
+        == AZ_OK);
 
-    az_span const result = builder;
+    az_span const result = az_span_slice(builder, 0, 10);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -498,11 +351,14 @@ static void test_az_span_replace(void** state)
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("12345678");
     az_span expected = AZ_SPAN_FROM_STR("XXX12345678");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 200 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 0, 0, AZ_SPAN_FROM_STR("XXX")) == AZ_OK);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 0, 0, AZ_SPAN_FROM_STR("XXX"))
+        == AZ_OK);
 
-    az_span const result = builder;
+    az_span const result = az_span_slice(builder, 0, 11);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -511,11 +367,14 @@ static void test_az_span_replace(void** state)
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("1");
     az_span expected = AZ_SPAN_FROM_STR("2");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 200 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 0, 1, AZ_SPAN_FROM_STR("2")) == AZ_OK);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 0, 1, AZ_SPAN_FROM_STR("2"))
+        == AZ_OK);
 
-    az_span const result = builder;
+    az_span const result = az_span_slice(builder, 0, 1);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -524,11 +383,14 @@ static void test_az_span_replace(void** state)
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("1234");
     az_span expected = AZ_SPAN_FROM_STR("4321");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 4 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 0, 4, AZ_SPAN_FROM_STR("4321")) == AZ_OK);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 0, 4, AZ_SPAN_FROM_STR("4321"))
+        == AZ_OK);
 
-    az_span const result = builder;
+    az_span const result = az_span_slice(builder, 0, 4);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -537,13 +399,19 @@ static void test_az_span_replace(void** state)
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("1234");
     az_span expected = AZ_SPAN_FROM_STR("1X34AB");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 10 - az_span_size(initial_state));
 
-    TEST_EXPECT_SUCCESS(_az_span_replace(&builder, 1, 2, AZ_SPAN_FROM_STR("X")));
+    TEST_EXPECT_SUCCESS(
+        _az_span_replace(builder, az_span_size(initial_state), 1, 2, AZ_SPAN_FROM_STR("X")));
 
-    TEST_EXPECT_SUCCESS(az_span_append(builder, AZ_SPAN_FROM_STR("AB"), &builder));
+    remainder = az_span_copy(remainder, AZ_SPAN_FROM_STR("AB"));
 
-    az_span const result = builder;
+    int32_t expected_remainder_size
+        = 10 - (az_span_size(AZ_SPAN_FROM_STR("AB")) + az_span_size(initial_state));
+    assert_int_equal(az_span_size(remainder), expected_remainder_size);
+
+    az_span const result = az_span_slice(builder, 0, 6);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -553,10 +421,13 @@ static void test_az_span_replace(void** state)
     az_span initial_state = AZ_SPAN_FROM_STR("123");
     az_span expected = AZ_SPAN_FROM_STR("1234");
 
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
-    TEST_EXPECT_SUCCESS(_az_span_replace(&builder, 3, 3, AZ_SPAN_FROM_STR("4")));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 4 - az_span_size(initial_state));
 
-    az_span const result = builder;
+    TEST_EXPECT_SUCCESS(
+        _az_span_replace(builder, az_span_size(initial_state), 3, 3, AZ_SPAN_FROM_STR("4")));
+
+    az_span const result = az_span_slice(builder, 0, 4);
     assert_true(az_span_is_content_equal(result, expected));
   }
   {
@@ -564,22 +435,25 @@ static void test_az_span_replace(void** state)
     uint8_t array[4];
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("1234");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 4 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 0, 4, AZ_SPAN_FROM_STR("4321X")) == AZ_ERROR_ARG);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 0, 4, AZ_SPAN_FROM_STR("4321X"))
+        == AZ_ERROR_ARG);
   }
   {
     // Fail on builder empty -> try to replace content from empty builder
     uint8_t array[200];
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
-    assert_true(_az_span_replace(&builder, 0, 1, AZ_SPAN_FROM_STR("2")) == AZ_ERROR_ARG);
+    assert_true(_az_span_replace(builder, 0, 0, 1, AZ_SPAN_FROM_STR("2")) == AZ_ERROR_ARG);
   }
   {
     // Replace content on empty builder -> insert at the end
     uint8_t array[200];
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
-    assert_true(_az_span_replace(&builder, 0, 0, AZ_SPAN_FROM_STR("2")) == AZ_OK);
-    az_span const result = builder;
+    assert_true(_az_span_replace(builder, 0, 0, 0, AZ_SPAN_FROM_STR("2")) == AZ_OK);
+    az_span const result = az_span_slice(builder, 0, 1);
     az_span const expected = AZ_SPAN_FROM_STR("2");
     assert_true(az_span_is_content_equal(result, expected));
   }
@@ -588,51 +462,302 @@ static void test_az_span_replace(void** state)
     uint8_t array[400];
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("1234");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 400 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 30, 31, AZ_SPAN_FROM_STR("4321X")) == AZ_ERROR_ARG);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 30, 31, AZ_SPAN_FROM_STR("4321X"))
+        == AZ_ERROR_ARG);
   }
   {
     // Fail when trying to replace out of bounds -> end position out
     uint8_t array[40];
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("1234");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 40 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 4, 5, AZ_SPAN_FROM_STR("4321X")) == AZ_ERROR_ARG);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 4, 5, AZ_SPAN_FROM_STR("4321X"))
+        == AZ_ERROR_ARG);
   }
   {
     // Fail when start is greater than end
     uint8_t array[40];
     az_span builder = AZ_SPAN_FROM_BUFFER(array);
     az_span initial_state = AZ_SPAN_FROM_STR("1234");
-    TEST_EXPECT_SUCCESS(az_span_append(builder, initial_state, &builder));
+    az_span remainder = az_span_copy(builder, initial_state);
+    assert_int_equal(az_span_size(remainder), 40 - az_span_size(initial_state));
 
-    assert_true(_az_span_replace(&builder, 3, 1, AZ_SPAN_FROM_STR("4321X")) == AZ_ERROR_ARG);
+    assert_true(
+        _az_span_replace(builder, az_span_size(initial_state), 3, 1, AZ_SPAN_FROM_STR("4321X"))
+        == AZ_ERROR_ARG);
   }
+}
+
+static void az_span_i64toa_test(void** state)
+{
+  (void)state;
+  uint8_t buffer[100];
+  az_span b_span = AZ_SPAN_FROM_BUFFER(buffer);
+  az_span remainder;
+  int32_t size_before_write = az_span_size(b_span);
+  int64_t number = 123;
+  az_span number_str = AZ_SPAN_FROM_STR("123");
+
+  assert_return_code(az_span_i64toa(b_span, number, &remainder), AZ_OK);
+  assert_int_equal(size_before_write, az_span_size(b_span));
+  // remainder should be size minus number of digits (3)
+  assert_int_equal(az_span_size(remainder), size_before_write - 3);
+
+  // create az_span for written data
+  b_span = az_span_init(az_span_ptr(b_span), az_span_size(b_span) - az_span_size(remainder));
+
+  assert_true(az_span_is_content_equal(b_span, number_str));
+
+  // convert back
+  uint64_t reverse = 0;
+  assert_return_code(az_span_atou64(b_span, &reverse), AZ_OK);
+  assert_int_equal(reverse, number);
+}
+
+static void az_span_i64toa_negative_number_test(void** state)
+{
+  (void)state;
+  uint8_t buffer[100];
+  az_span b_span = AZ_SPAN_FROM_BUFFER(buffer);
+  az_span remainder;
+  int32_t size_before_write = az_span_size(b_span);
+  int64_t number = -123;
+  az_span number_str = AZ_SPAN_FROM_STR("-123");
+
+  assert_return_code(az_span_i64toa(b_span, number, &remainder), AZ_OK);
+  assert_int_equal(size_before_write, az_span_size(b_span));
+  // remainder should be size minus number of digits (4)
+  assert_int_equal(az_span_size(remainder), size_before_write - 4);
+
+  // create az_span for written data
+  b_span = az_span_init(az_span_ptr(b_span), az_span_size(b_span) - az_span_size(remainder));
+
+  assert_true(az_span_is_content_equal(b_span, number_str));
+
+  // convert back TODO: az_span_ato64 should support negative numbers since az_span_i64toa support
+  // it. https://github.com/Azure/azure-sdk-for-c/issues/598
+  /* uint64_t reverse = 0;
+  assert_return_code(az_span_atou64(b_span, &reverse), AZ_OK);
+  assert_int_equal(reverse, number);
+  */
+}
+
+static void az_span_slice_to_end_test(void** state)
+{
+  (void)state;
+  uint8_t raw_buffer[20];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+
+  assert_int_equal(az_span_size(buffer), 20);
+
+  az_span result = az_span_slice_to_end(buffer, 1);
+  assert_int_equal(az_span_size(result), 19);
+
+  result = az_span_slice_to_end(buffer, 5);
+  assert_int_equal(az_span_size(result), 15);
+}
+
+static void az_span_test_macro_only_allows_byte_buffers(void** state)
+{
+  (void)state;
+  {
+    uint8_t uint8_buffer[2];
+    assert_int_equal(_az_IS_ARRAY(uint8_buffer), 1);
+    assert_int_equal(_az_IS_BYTE_ARRAY(uint8_buffer), 1);
+    az_span valid = AZ_SPAN_FROM_BUFFER(uint8_buffer);
+    assert_int_equal(az_span_size(valid), 2);
+  }
+
+  {
+    char char_buffer[2];
+    assert_int_equal(_az_IS_ARRAY(char_buffer), 1);
+    assert_int_equal(_az_IS_BYTE_ARRAY(char_buffer), 1);
+    az_span valid = AZ_SPAN_FROM_BUFFER(char_buffer);
+    assert_int_equal(az_span_size(valid), 2);
+  }
+
+  {
+    uint32_t uint32_buffer[2];
+    assert_int_equal(_az_IS_ARRAY(uint32_buffer), 1);
+    assert_int_equal(_az_IS_BYTE_ARRAY(uint32_buffer), 0);
+  }
+
+  {
+    uint8_t x = 1;
+    uint8_t* p1 = &x;
+    assert_int_equal(_az_IS_ARRAY(p1), 0);
+    assert_int_equal(_az_IS_BYTE_ARRAY(p1), 0);
+  }
+
+  {
+    char* p1 = "HELLO";
+    assert_int_equal(_az_IS_ARRAY(p1), 0);
+    assert_int_equal(_az_IS_BYTE_ARRAY(p1), 0);
+  }
+}
+
+static void az_span_from_str_succeeds(void** state)
+{
+  (void)state;
+  char* str = "HelloWorld";
+  az_span buffer = az_span_from_str(str);
+
+  assert_int_equal(az_span_size(buffer), 10);
+  assert_true(az_span_ptr(buffer) != NULL);
+  assert_true((char*)az_span_ptr(buffer) == str);
+}
+
+static void az_span_copy_uint8_succeeds(void** state)
+{
+  (void)state;
+  uint8_t raw_buffer[15];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+
+  buffer = az_span_copy_u8(buffer, 'a');
+  assert_int_equal(az_span_size(buffer), 14);
+  buffer = az_span_copy_u8(buffer, 'b');
+  assert_int_equal(az_span_size(buffer), 13);
+  buffer = az_span_copy_u8(buffer, 'c');
+  assert_int_equal(az_span_size(buffer), 12);
+
+  assert_true(az_span_is_content_equal(
+      az_span_slice(AZ_SPAN_FROM_BUFFER(raw_buffer), 0, 3), AZ_SPAN_FROM_STR("abc")));
+}
+
+static void az_span_i32toa_succeeds(void** state)
+{
+  (void)state;
+  int32_t v = 12345;
+  uint8_t raw_buffer[15];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+  az_span out_span;
+
+  assert_true(az_succeeded(az_span_i32toa(buffer, v, &out_span)));
+  assert_int_equal(az_span_size(out_span), 10);
+  assert_true(az_span_is_content_equal(
+      az_span_slice(AZ_SPAN_FROM_BUFFER(raw_buffer), 0, 5), AZ_SPAN_FROM_STR("12345")));
+}
+
+static void az_span_i32toa_negative_succeeds(void** state)
+{
+  (void)state;
+  int32_t v = -12345;
+  uint8_t raw_buffer[15];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+  az_span out_span;
+
+  assert_true(az_succeeded(az_span_i32toa(buffer, v, &out_span)));
+  assert_int_equal(az_span_size(out_span), 9);
+  assert_true(az_span_is_content_equal(
+      az_span_slice(AZ_SPAN_FROM_BUFFER(raw_buffer), 0, 6), AZ_SPAN_FROM_STR("-12345")));
+}
+
+static void az_span_i32toa_zero_succeeds(void** state)
+{
+  (void)state;
+  int32_t v = 0;
+  uint8_t raw_buffer[15];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+  az_span out_span;
+
+  assert_true(az_succeeded(az_span_i32toa(buffer, v, &out_span)));
+  assert_int_equal(az_span_size(out_span), 14);
+  assert_true(az_span_is_content_equal(
+      az_span_slice(AZ_SPAN_FROM_BUFFER(raw_buffer), 0, 1), AZ_SPAN_FROM_STR("0")));
+}
+
+static void az_span_i32toa_max_int_succeeds(void** state)
+{
+  (void)state;
+  int32_t v = 2147483647;
+  uint8_t raw_buffer[15];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+  az_span out_span;
+
+  assert_true(az_succeeded(az_span_i32toa(buffer, v, &out_span)));
+  assert_int_equal(az_span_size(out_span), 5);
+  assert_true(az_span_is_content_equal(
+      az_span_slice(AZ_SPAN_FROM_BUFFER(raw_buffer), 0, 10), AZ_SPAN_FROM_STR("2147483647")));
+}
+
+static void az_span_i32toa_overflow_fails(void** state)
+{
+  (void)state;
+  int32_t v = 2147483647;
+  uint8_t raw_buffer[4];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+  az_span out_span;
+
+  assert_true(az_span_i32toa(buffer, v, &out_span) == AZ_ERROR_INSUFFICIENT_SPAN_SIZE);
+}
+
+static void az_span_u32toa_succeeds(void** state)
+{
+  (void)state;
+  uint32_t v = 12345;
+  uint8_t raw_buffer[15];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+  az_span out_span;
+
+  assert_true(az_succeeded(az_span_u32toa(buffer, v, &out_span)));
+  assert_int_equal(az_span_size(out_span), 10);
+  assert_true(az_span_is_content_equal(
+      az_span_slice(AZ_SPAN_FROM_BUFFER(raw_buffer), 0, 5), AZ_SPAN_FROM_STR("12345")));
+}
+
+static void az_span_u32toa_zero_succeeds(void** state)
+{
+  (void)state;
+  uint32_t v = 0;
+  uint8_t raw_buffer[15];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+  az_span out_span;
+
+  assert_true(az_succeeded(az_span_u32toa(buffer, v, &out_span)));
+  assert_int_equal(az_span_size(out_span), 14);
+  assert_true(az_span_is_content_equal(
+      az_span_slice(AZ_SPAN_FROM_BUFFER(raw_buffer), 0, 1), AZ_SPAN_FROM_STR("0")));
+}
+
+static void az_span_u32toa_max_uint_succeeds(void** state)
+{
+  (void)state;
+  uint32_t v = 4294967295;
+  uint8_t raw_buffer[15];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+  az_span out_span;
+
+  assert_true(az_succeeded(az_span_u32toa(buffer, v, &out_span)));
+  assert_int_equal(az_span_size(out_span), 5);
+  assert_true(az_span_is_content_equal(
+      az_span_slice(AZ_SPAN_FROM_BUFFER(raw_buffer), 0, 10), AZ_SPAN_FROM_STR("4294967295")));
+}
+
+static void az_span_u32toa_overflow_fails(void** state)
+{
+  (void)state;
+  uint32_t v = 2147483647;
+  uint8_t raw_buffer[4];
+  az_span buffer = AZ_SPAN_FROM_BUFFER(raw_buffer);
+  az_span out_span;
+
+  assert_true(az_span_u32toa(buffer, v, &out_span) == AZ_ERROR_INSUFFICIENT_SPAN_SIZE);
 }
 
 int test_az_span()
 {
   const struct CMUnitTest tests[] = {
+    cmocka_unit_test(az_span_slice_to_end_test),
     cmocka_unit_test(test_az_span_getters),
     cmocka_unit_test(az_single_char_ascii_lower_test),
-    cmocka_unit_test(az_span_append_uint8_overflow_fails),
-    cmocka_unit_test(az_span_append_uint8_succeeds),
-    cmocka_unit_test(az_span_append_i32toa_succeeds),
-    cmocka_unit_test(az_span_append_i32toa_negative_succeeds),
-    cmocka_unit_test(az_span_append_i32toa_zero_succeeds),
-    cmocka_unit_test(az_span_append_i32toa_max_int_succeeds),
-    cmocka_unit_test(az_span_append_i32toa_NULL_span_fails),
-    cmocka_unit_test(az_span_append_i32toa_overflow_fails),
-    cmocka_unit_test(az_span_append_u32toa_succeeds),
-    cmocka_unit_test(az_span_append_u32toa_zero_succeeds),
-    cmocka_unit_test(az_span_append_u32toa_max_uint_succeeds),
-    cmocka_unit_test(az_span_append_u32toa_NULL_span_fails),
-    cmocka_unit_test(az_span_append_u32toa_overflow_fails),
     cmocka_unit_test(az_span_to_lower_test),
-    cmocka_unit_test(az_span_to_uint64_return_errors),
-    cmocka_unit_test(az_span_to_uint32_test),
     cmocka_unit_test(az_span_to_str_test),
     cmocka_unit_test(az_span_find_beginning_success),
     cmocka_unit_test(az_span_find_middle_success),
@@ -648,6 +773,22 @@ int test_az_span()
     cmocka_unit_test(az_span_find_capacity_checks_success),
     cmocka_unit_test(az_span_find_overlapping_checks_success),
     cmocka_unit_test(test_az_span_replace),
+    cmocka_unit_test(az_span_atou64_return_errors),
+    cmocka_unit_test(az_span_atou32_test),
+    cmocka_unit_test(az_span_i64toa_negative_number_test),
+    cmocka_unit_test(az_span_i64toa_test),
+    cmocka_unit_test(az_span_test_macro_only_allows_byte_buffers),
+    cmocka_unit_test(az_span_from_str_succeeds),
+    cmocka_unit_test(az_span_copy_uint8_succeeds),
+    cmocka_unit_test(az_span_i32toa_succeeds),
+    cmocka_unit_test(az_span_i32toa_negative_succeeds),
+    cmocka_unit_test(az_span_i32toa_max_int_succeeds),
+    cmocka_unit_test(az_span_i32toa_zero_succeeds),
+    cmocka_unit_test(az_span_i32toa_overflow_fails),
+    cmocka_unit_test(az_span_u32toa_succeeds),
+    cmocka_unit_test(az_span_u32toa_zero_succeeds),
+    cmocka_unit_test(az_span_u32toa_max_uint_succeeds),
+    cmocka_unit_test(az_span_u32toa_overflow_fails),
   };
   return cmocka_run_group_tests_name("az_core_span", tests, NULL, NULL);
 }
